// packages/react-signer/src/Checks/types.ts
// Copyright 2017-2020 @polkadot/react-signer authors & contributors
// This software may be modified and distributed under the terms
// of the Apache-2.0 license. See the LICENSE file for details.

import BN from 'bn.js';

export interface ExtraFees {
  extraAmount: BN;
  extraFees: BN;
  extraWarn: boolean;
}